Straighten out the optional view integration
