--- spec/isolation/isolation_spec_helper.rb
@@ -1,18 +1,0 @@
-# frozen_string_literal: true
-
-require "rubygems"
-require "bundler"
-Bundler.setup(:default, :development, :test)
-
-$LOAD_PATH.unshift "lib"
-require "rspec"
-
-module RSpec
-  module Support
-    module Runner
-      def self.run
-        Core::Runner.autorun
-      end
-    end
-  end
-end
